Complete first 4 exercises
Exercise 4 uses a less efficient and more complicated way of
removing the elements from the array but is what I was able to
come up with before looking at the solution. After looking at the
solution, it's clear it would've been simpler to just return a new
array that contained everything from the old array *except* for
the elements to be removed.