Complete exercise 5
Get the sum of all number from and including the arguments passed
to the function. A little sloppier than the solution given. typeof
is redundant in the first if statement. first probably shouldn't be
the value that changes in the for loop (more like let i = first). It
works fine as is but would be more readable.