Complete JavaScript Exercises 09 and 10
In exercise 09, I made the input string all lowercase in order
to be case insensitive. I then split the string so I could then
filter it with the formatStr function which simply checks the
given character against the regex and returns true if the
character is not whitespace or punctuation. The array gets
rejoined into a string, now all lowercase and without any
puncuation or whitespace. It then gets sliced into firstHalf
and lastHalf. lastHalf gets reversed so that it (if the input is
in fact a palindrome) matches firstHalf.

In exercise 10, the function returns the nth fibonacci number.
This is done with a simple for loop that adds the previous two
numbers in the sequence.