Compelete JavaScript Exercise 12
This is the final JavaScript exercise of the Foundations Course.
I used the reduce method to find the oldest person out of an array
of people objects. The calculateAge helper functions can find the
person's age either by subtracting the yearOfBirth from the
yearOfDeath or by subtracting the yearOfBirth from the current year
if there is no given yearOfDeath. The reduce function returns the
currentPerson if currentAge is greater than oldestAge and oldest
otherwise.